+++ src/components/sidebar/Sidebar.tsx
@@ -24,7 +24,7 @@ import * as s from './sidebar.scss';
 // This is due to blocked updates issue
 // tslint:disable-next-line
 // https://github.com/ReactTraining/react-router/blob/master/packages/react-router/docs/guides/blocked-updates.md
-interface ISidebarProps{
+interface ISidebarProps {
     routeStore?: RouteStore;
     searchStore?: SearchStore;
     toolbarStore?: ToolbarStore;
