+++ src/components/sidebar/routesView/routeList.scss
@@ -1,5 +1,5 @@
 @import '../../shared/styles/common.scss';
-.routesListView {
+.routeListView {
     display: flex;
     flex-direction: column;
     overflow-y: auto;
